requirements.txt
fastapi==0.111.0
uvicorn==0.30.1

scikit-learn==1.5.0
numpy==1.26.4
pandas==2.2.2
joblib==1.4.2

psycopg2-binary==2.9.9

openai==1.35.13
httpx==0.27.0

pydantic==2.7.4